fix: Use package resources instead of data files (#642)
## Summary

- Including project scaffolding directly in the cli package so that
`zipline init` can port between wheel and editable mode. This makes it
easier to reason about across platforms as well, since everything is
bundled in the wheel.


## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update
<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Added support for specifying a cloud provider when initializing, with
options for AWS or GCP.
- Introduced example data sources and group-by configurations for
purchase data aggregation on GCP.
- Added predefined team configurations for GCP environments, including
default, test, and detailed configuration options.

- **Chores**
- Updated package setup to refine which packages and resources are
included in the distribution.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->

---------

Co-authored-by: Thomas Chow <thomaschow369@gmail.com>