fix: get the full table name properly from Driver (#641)
## Summary

- #586 did a refactor on the naming of internal tables - we need to
update the Driver code to account for this.

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update
<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Bug Fixes**
- Improved handling of table name parsing by changing the delimiter from
a dot to a double underscore when extracting namespace and table name.
- Updated the method for computing source output table names to enhance
accuracy in table identification.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->

Co-authored-by: Thomas Chow <thomaschow369@gmail.com>