chore: Releases Reference Specific Jar Version (#647)
## Summary

This change removes our reliance on the "passing-candidate" version and
instead uses the last release version + the current sha of
main-passing-tests to pull the correct version of the jars from the
repos.

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [x ] Integration tested
- [ ] Documentation update



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Chores**
- Improved artifact handling in release workflows to fetch and promote
jar files based on the previous release version and latest commit.
- Removed intermediate artifact push steps for "passing-candidate"
releases in canary workflows, streamlining the process.
- Updated workflow dependencies to clean up artifacts immediately after
integration tests.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->