Revert "chore: Releases Reference Specific Jar Version" (#663)
Reverts zipline-ai/chronon#647

There is a race condition for when a release is cut while the
integration tests are running. This leads to a different version
identifier than the expected one as the integration tests use the latest
release number which can be the one that was cut and failed.