chore: remove dep on Extensions from TableUtils (#682)
## Summary

- TableUtils shouldn't need a dep on Extensions. It's the other way
around. We'll just clone the `effectivePartitionColumn` logic here. Not
ideal, but it provides a much more streamlined dep structure.

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update
<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Refactor**
- Improved internal handling of query parameters for more concise and
maintainable code. No changes to user-facing functionality.
- **Chores**
- Updated build configuration by removing an unused dependency and
refining dependency specifications.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->

---------

Co-authored-by: Thomas Chow <thomaschow369@gmail.com>