feat: do partition filtering on bq native tables by union individual partitions (#690)
## Summary

- Getting a 403 querying for a range of partitions in bigquery native
tables:
```
Response too large to return. Consider specifying a destination table in your job configuration
```
- instead, let's just query individual partitions of data as separate
dataframes and union them together.

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update



<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Bug Fixes**
- Improved handling of BigQuery partitioned tables, ensuring more
accurate partition filtering and data retrieval.

- **Refactor**
- Streamlined the process for reading partitioned data from BigQuery,
resulting in a clearer and more consistent approach for users working
with partitioned tables.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Thomas Chow <thomaschow369@gmail.com>