feat: still use information_schema for bigquery native tables (#710)
## Summary

- In BigQuery, we have views and tables that can be native. For native
tables we can partition list through the information schema. We cannot
do the same for views. We should take two different approaches for
partition listing for tables and views.In order to do this, we'll do a
blind test - first check the information schema and if we can't get a
partition column out of that, we'll just do a blind `select
distinct(...)`.

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update

<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Bug Fixes**
- Improved error handling for missing partition columns, providing
clearer error messages and a more robust fallback method for retrieving
partition values.
  
- **Refactor**
- Centralized the handling of missing partition columns for more
consistent behavior across the application.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Thomas Chow <thomaschow369@gmail.com>