Update to a different metadata path in GCS everytime to avoid collisions (#771)
## Summary

^^^

Currently, we'll face unexpected behavior if multiple people are working
and iterating on the same GroupBy/Join and changing the conf because
we'll upload to the same GCS path.

This change adds the job id to the destination GCS path.

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [x] Integration tested
- [ ] Documentation update



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Refactor**
- Streamlined job submission to upload a single metadata configuration
file, simplifying the process.
- Enhanced job ID management by requiring and propagating a job ID
argument, improving job tracking and consistency.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->