fix: run setups regardless of partitioning (#777)
## Summary

- We should be running setups regardless of whether things are
partitioned.

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update


<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Refactor**
- Adjusted the timing of SQL setup command execution to occur earlier in
the staging query process, ensuring setups run before any query
execution or partition checks. No changes to user-facing features or
functionality.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

Co-authored-by: thomaschow <thomaschow369@gmail.com>