Remove cloud gcp embedded target + pull out 2.13 builds (#811)
## Summary
Pull out the embedded cloud gcp jar approach as its not needed anymore.
Yanked out the target + publish steps. Also removed the Scala 2.13 build
pieces as we also don't need that anymore (it was only for embedded).
This will help speed up CI and simplify our dev flow as we don't need to
worry about supporting 2.12 and 2.13 atm. When the time comes we can
switch over to 2.13 across the board when the underlying engines / cloud
services support it.

## Checklist
- [ ] Added Unit Tests
- [X] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Chores**
- Removed all Scala 2.13 build and test support, including related
configuration files, dependencies, and CI workflows.
- Deleted the embedded GCP jar build and deployment steps, as well as
associated artifacts.
  - Updated documentation to reflect removal of Scala 2.13 instructions.
- Unconditionally enabled code formatting for all Scala targets,
regardless of Scala version.
- Cleaned up build rules and dependencies to remove references to Scala
2.13 and obsolete artifacts.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->