Refactor + add support for entities in Flink's AvroCodecFn (#842)
## Summary
The AvroCodecFn in Flink handles converting tiles / ETLed events to avro
objects ready for persisting to the KV. The version we've had so far
just supports event based GroupBys. This PR extends that to also support
entity based GroupBys (in the untiled implementation). Refactored some
code in SparkExprEval to help with some reuse.

I yanked out the key null checks I added in the previous PR
(https://github.com/zipline-ai/chronon/pull/833). If a user's key is
seemingly based on a column that has an explode transform it results in
a Spark error as we can't have an explode in the where clause. We didn't
have this prior so we can dig into adding it as a follow up.

## Checklist
- [X] Added Unit Tests
- [X] Covered by existing CI
- [X] Integration tested - Launched a streaming job in canary and
confirmed it came up & wrote data to BT
- [ ] Documentation update



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **New Features**
- Added integration tests to validate end-to-end processing for
entity-based group-by operations in Flink jobs.
- Enhanced support for additional data model types in Flink processing.

- **Refactor**
- Improved flexibility in determining key, value, and event time columns
for data processing.
- Generalized event source utilities to support multiple event types
with custom watermark strategies.
  - Updated class and method names for clarity and consistency.

- **Tests**
- Introduced new and renamed integration test classes to improve test
coverage and organization.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->