feat: disable ntz timestamp inference (#855)
## Summary

- We don't support TimestampNTZ spark types in the avro conversions for
chronon's type system. Let's default this to the community desired
setting in spark 4.0: https://issues.apache.org/jira/browse/SPARK-47493

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update


<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":""}
```
-->

Co-authored-by: thomaschow <thomaschow369@gmail.com>