Update avro Row conversions to handle logical time type + tests (#865)
## Summary
When users include expressions like `unix_millis(TIMESTAMP(ts))` we seem
to be ending up with seconds resolution in the streaming jobs instead of
millis res. This seems to be due to us treating the timestamp as millis
resolution and then calling unix_millis on it (which divs by 1000).
Instead the Spark type should be of micros resolution which is what our
conversion code should vend out. Updated our avro CU tests to cover this
case (they fail now if we revert the Row update) and also refactored the
tests a bit to reuse some of the streaming deser related classes rather
than the old AvroDataToCatalyst way which we got away from after the
switch to Mutations.

## Checklist
- [X] Added Unit Tests
- [X] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Bug Fixes**
- Improved timestamp handling by updating Avro schema definitions and
conversions to use microseconds precision instead of milliseconds.
- Fixed test assertions to verify timestamp fields with a fixed value
for consistency.

- **Chores**
- Updated dependencies and configurations to enable Java 8 DateTime API
and adjust Parquet timestamp inference settings.
- Removed unused Avro-to-Catalyst utility code to streamline the
codebase.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->