Start emitting Flink processing time metric (#874)
## Summary
Currently we track the feature freshness (event created -> sink). In
some cases there's a sizeable lag upstream so it's useful to also track
the processing time within Flink as a metric of its own. To emit this
metric though we need to plumb the start processing time field from the
source (just prior to when we deser the event) and have it flow through
the operators to the sink. The bulk of this PR is really doing that
plumbing and wiring up of things. Should make emitting future metrics
from source -> sink easier.

Tested on our canary cluster and confirmed the metrics show up -
![Screenshot 2025-06-18 at 11 56
55 AM](https://github.com/user-attachments/assets/c424ab70-1b57-49f1-bfec-d4d77f1a20be)


## Checklist
- [ ] Added Unit Tests
- [X] Covered by existing CI
- [X] Integration tested
- [ ] Documentation update



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **New Features**
- Added tracking of event processing start time throughout the Flink
data pipeline, enabling more detailed monitoring and metrics.
  - Introduced a new histogram metric to monitor Flink processing times.

- **Improvements**
- Enhanced event handling by encapsulating event data and processing
metadata in a unified structure for more robust processing and
observability.

- **Bug Fixes**
- Updated tests to ensure compatibility with new event structure and
timestamp tracking.

- **Chores**
- Updated ignored directories to prevent unnecessary files from being
tracked.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->