Wire up optional debugging logging in Flink (#876)
## Summary
While debugging issues in a GroupBy streaming job dealing with lastK
features over Spark structs, we found it very helpful to be able to log
more finely what was being done in the Flink app per event. This logging
is verbose and might contain info we don't want to log and is thus
turned off by default. A user can enable and turn this on a job-by-job
basis (with a `--enable-debug` flag) to aid in debugging.

Tested manually to confirm the default behavior is logging off and when
we enable we do see additional logs.

## Checklist
- [ ] Added Unit Tests
- [X] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **New Features**
- Added a new debug mode flag that enables detailed logging for various
components in both Python and Scala modules.
- Introduced a command-line option and job flag to toggle verbose debug
logging for supported run modes and Flink jobs.

- **Improvements**
- Enhanced logging control, allowing users to enable or disable debug
information for easier troubleshooting and diagnostics.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->