Update GcpApiImpl to create single BT client and pools (#913)
## Summary
While debugging high BT write latencies from a Flink job we noticed that
there were a ton of BigTable related threads being created. The Flink
task manager had ~5K threads (of which maybe 4-4.5K were
chronon-bt-gax-* and gax-*). The issue turns out to be that in case the
task restarts due to an uncaught exception (in this case there are no
messages on pubsub so the pubsub connector throws on timeout), Flink
restarts the task but not the JVM. This results in the `.open(..)`
method being called on our various operators and in the
AsyncKVStoreWriter we can genKVStore on the API (GcpApiImpl). This
results in the BT client and thread pools being recreated.

We also want to share the client and threadpools across the tasks in the
task manager JVM instead of creating n of these (n == number of task
slots per task manager == 4 currently). This PR tackles both of these.

Was able to confirm with a running job in canary that - we have only 64
bt-gax threads (16 cores * 4) and this number doesn't grow when I
trigger task restarts (by shutting down the pubsub event driver).

## Checklist
- [ ] Added Unit Tests
- [X] Covered by existing CI
- [X] Integration tested
- [ ] Documentation update



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

* **Refactor**
* Improved performance by ensuring the key-value store instance is
shared and initialized only once per process, reducing resource usage in
multi-task environments.
* **Other**
* Added logging to provide more visibility when a new key-value store
instance is created.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->